feat: Touhou Little Maid: Tsumugi — Minecraft 1.21.11 Fabric port
Independently maintained community fork continuing the Fabric port lineage
(TartaricAcid's original Touhou Little Maid, Sh1roCu's Orihime) to Minecraft 1.21.11.

Development source: 77f0480e1b4f0c549ad010cb5e865321cf33afa9
Verified on the clean public tree by a full build (BUILD SUCCESSFUL), JUnit 38/38,
non-comment token parity with the development source (mismatches=0), and git diff --check.
The token-identical development source additionally passed GameTest 48/48 without optional
mods and 48/48 with Kaleidoscope Cookery, Kaleidoscope Tavern and Farmer's Delight installed.

+++ src/main/java/cn/sh1rocu/touhoulittlemaid/mixin/common/FarmBlockMixin.java
@@ -14,8 +14,9 @@ import org.spongepowered.asm.mixin.injection.At;
 
 @Mixin(FarmBlock.class)
 public class FarmBlockMixin {
+
     @WrapWithCondition(method = "fallOn", at = @At(value = "INVOKE", target = "Lnet/minecraft/world/level/block/FarmBlock;turnToDirt(Lnet/minecraft/world/entity/Entity;Lnet/minecraft/world/level/block/state/BlockState;Lnet/minecraft/world/level/Level;Lnet/minecraft/core/BlockPos;)V"))
-    private boolean tlm$onFarmlandTrample(Entity entity, BlockState state, Level level, BlockPos pos, @Local(argsOnly = true) float fallDistance) {
-        return CommonHooks.onFarmlandTrample(level, pos, Blocks.DIRT.defaultBlockState(), fallDistance, entity);
+    private boolean tlm$onFarmlandTrample(Entity entity, BlockState state, Level level, BlockPos pos, @Local(argsOnly = true) double fallDistance) {
+        return CommonHooks.onFarmlandTrample(level, pos, Blocks.DIRT.defaultBlockState(), (float) fallDistance, entity);
     }
 }
