+++ src/main/java/com/github/tartaricacid/touhoulittlemaid/geckolib3/core/keyframe/bone/LinearKeyFrame.java
@@ -8,14 +8,14 @@ public class LinearKeyFrame extends BoneKeyFrame {
     private final Vector3v endPoint;
     private final Vector3v postPoint;
 
-    public LinearKeyFrame(double startTick, double totalTick, Vector3v beginPoint, Vector3v endPoint, Vector3v postPoint) {
+    public LinearKeyFrame(float startTick, float totalTick, Vector3v beginPoint, Vector3v endPoint, Vector3v postPoint) {
         super(startTick, totalTick, beginPoint);
         this.endPoint = endPoint;
         this.postPoint = postPoint;
     }
 
     @Override
-    public Vector3f getLerpPoint(ExpressionEvaluator<?> evaluator, double percentCompleted) {
+    public Vector3f getLerpPoint(ExpressionEvaluator<?> evaluator, float percentCompleted) {
         if (isBegin(percentCompleted)) {
             return beginPoint.eval(evaluator);
         }
